% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calib.stat}
\alias{calib.stat}
\title{calib.stat}
\usage{
calib.stat(p, T.test, C.test, cutpts, t)
}
\arguments{
\item{p}{Predicted event probablities}

\item{T.test}{Observation times from the test set}

\item{C.test}{Event indicators from the test set}

\item{cutpts}{Vector of cutpoints defining risk categories}

\item{t}{Time horizon of predictions}
}
\value{
The calibration statistic
}
\description{
Calibration statistic for censored time-to-event data
}
\details{
Compute the calibration statistic for censored time-to-event data
}

