% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calib.table}
\alias{calib.table}
\title{calib.table}
\usage{
calib.table(LP, T.test, C.test, cutpts, t, modelnames = rep("", length(LP)))
}
\arguments{
\item{LP}{List of vectors of predicted event probablities.}

\item{T.test}{Observation times from the test set.}

\item{C.test}{Event indicators from the test set.}

\item{cutpts}{Vector of cutpoints defining risk categories.}

\item{t}{Time horizon of predictions.}

\item{modelnames}{(Optional) Names of the models in \code{LP}, for pretty printing.}
}
\value{
A calibration table, with one row for each category defined by \code{cutpts}.
}
\description{
Calibration table for censored time-to-event data
}
\details{
Compute the calibration table for censored time-to-event data
}

