% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcbinom.R
\name{dcnbinom}
\alias{dcnbinom}
\title{contaminated Negative Binomial density}
\usage{
dcnbinom(x, mu, alpha, delta, eta, log = F)
}
\arguments{
\item{x}{vector of (non negative integer) quantities.}

\item{mu}{mean parameter. mu > 0.}

\item{alpha}{dispersion parameter. alpha > 0.}

\item{delta}{proportion of mild outliers (bad points). 0 < delta < 1.}

\item{eta}{inflation parameter. eta > 1.}

\item{log}{loglical; if True probabilities p are given as log(p). Default is False.}
}
\value{
The density of the contaminated negative binomial distribution.
}
\description{
Density function of contaminated negative binomial distribution
}
\details{
The contaminated negative binomial distribution with mu=\eqn{\mu}, alpha=\eqn{\alpha}, delta=\eqn{\delta}, and eta=\eqn{\eta} has density
\deqn{
f(x;\mu,\alpha,\delta,\eta)=(1-\delta)\binom{x-1/\alpha-1}{x}(\frac{\alpha\mu}{1+\alpha\mu})^x(\frac{1}{1+\alpha\mu})^{1/\alpha}+\delta\binom{x-1/(\eta\alpha)-1}{x}(\frac{\eta\alpha\mu}{1+\eta\alpha\mu})^x(\frac{1}{1+\eta\alpha\mu})^{1/(\eta\alpha)}
}
for, \eqn{x=0,1,\dots}, \eqn{\mu>0, \alpha>0, 0<\delta<1} and \eqn{\eta>1}.
}
