% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcbinom_regression.R
\name{rcnbinom_regression}
\alias{rcnbinom_regression}
\title{Generate Contaminated Negative Binomial Regression Data}
\usage{
rcnbinom_regression(
  n = 50000,
  off = 0,
  alpha = 0.5,
  delta = 0.95,
  eta = 2,
  xv = c(1, 0.75, -1.5)
)
}
\arguments{
\item{n}{Number of observations to generate. Default is 50000.}

\item{off}{Offset term. Default is 0.}

\item{alpha}{Dispersion parameter. Default is 0.5.}

\item{delta}{Proportion of good data points. Default is 0.95.}

\item{eta}{Inflation parameter for bad data points. Default is 2.}

\item{xv}{Vector of regression coefficients. Default is c(1, 0.75, -1.5).}
}
\value{
A data frame containing the generated contaminated negative binomial response variable and the predictor variables.
}
\description{
This function generates synthetic data for a contaminated negative binomial regression model.
}
