% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcbinom_regression.R
\name{rnbinom_regression}
\alias{rnbinom_regression}
\title{Generate Negative Binomial Regression Data}
\usage{
rnbinom_regression(n = 1, off = 0, alpha = 0.5, xv = c(1, 0.75, -1.5))
}
\arguments{
\item{n}{Number of observations to generate. Default is 1.}

\item{off}{Offset term. Default is 0.}

\item{alpha}{Dispersion parameter. Default is 0.5.}

\item{xv}{Vector of regression coefficients. Default is c(1, 0.75, -1.5).}
}
\value{
A data frame containing the generated negative binomial response variable and the predictor variables.
}
\description{
This function generates synthetic data for a negative binomial regression model.
}
